pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aou-workbench"
version = "0.1.0"
description = "Source-of-truth All of Us genomics workbench with rhabdomyolysis analysis workflows."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "scipy>=1.10",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
workbench = [
  "google-cloud-bigquery>=3.25",
  "hail>=0.2",
]

[project.scripts]
aou-workbench = "aou_workbench.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
